Update multi select examples & nightwatch tests.
