+++ packages/terra-list/tests/nightwatch/multi-select-list/DefaultMultiSelectList.jsx
@@ -2,4 +2,12 @@ import React from 'react';
 
 import MultiSelectList from '../../../lib/MultiSelectList';
 
-export default () => <MultiSelectList />;
+const list = () => (
+  <MultiSelectList>
+    <MultiSelectList.Item content={<p>test 1</p>} key="123" isSelectable />
+    <MultiSelectList.Item content={<p>test 2</p>} key="124" isSelectable />
+    <MultiSelectList.Item content={<p>test 3</p>} key="125" isSelectable />
+  </MultiSelectList>
+ );
+
+export default list;
